Move functions to tools. Write function to export Excel sheet to CSV. Bug exists: need to use destination file.
