+++ Analysis/python/tools.py
@@ -4,8 +4,10 @@ import csv
 import os
 import sys
 import shutil
+import datetime
 import math
 import numpy as np
+import pandas as pd
 
 # general error code
 ERROR_CODE = -999
@@ -17,6 +19,9 @@ def makeDir(dir_name):
 
 # move file if it exists; if not, print error and exit
 def moveFile(source_file, destination_file):
+    print("Moving Excel file...")
+    print(f" - source file: {source_file}")
+    print(f" - destination file: {destination_file}")
     try:
         shutil.move(source_file, destination_file)
     except FileNotFoundError:
@@ -29,6 +34,41 @@ def appendSlash(path):
     if path[-1] != slash:
         path += slash
     return path
+
+# get today's date
+def getTodayDate():
+    date_format = "%Y-%m-%d"
+    today_date_object = datetime.datetime.today()
+    today_date = today_date_object.strftime(date_format)
+    return today_date
+
+# add date to file name
+def addDateToFileName(original_file_name):
+    print("Adding date to file name...")
+
+    today_date = getTodayDate()
+    today_date = today_date.replace("-", "_")
+    file_base, file_extension = os.path.splitext(original_file_name)
+    new_file_name = f"{file_base}_{today_date}{file_extension}"
+    
+    print(f" - today's date: {today_date}")
+    print(f" - original file name: {original_file_name}")
+    print(f" - new file name: {new_file_name}")
+    
+    return new_file_name
+
+# export Excel sheet to CSV
+def exportExcelSheetToCSV(excel_file, excel_sheet):
+    print("Exporting Excel sheet to CSV...")
+    print(f" - Excel file: {excel_file}")
+    print(f" - Excel sheet: {excel_sheet}")
+    try:
+        df = pd.read_excel(excel_file, sheet_name=excel_sheet)
+        base, extension = os.path.splitext(excel_file)
+        csv_file = base + '.csv'
+        df.to_csv(csv_file, index=False)
+    except Exception as e:
+        print(f"ERROR: {e}")
 
 # check if string can be converted to int
 def is_int(num):
